Add  nil default to input values and expose prefix
Pulumi config secret values in json output will not have a 'value' key so we set it to nil by default.

A future feature could be an option to expose secret inputs to InSpec controls without relying on '$ pulumi config get' which would expose the secret in plaintext in the kitchen log.